+++ CHANGES.rst
@@ -2,6 +2,14 @@
  change log for taxtastic
 ==========================
 
+0.4
+===
+
+ * 'names' table in the taxonomy database has a new column
+   'is_classified' indicating whether 'tax_name' should be considered
+   "classified".
+
+
 0.3.2
 =====